shared: clientside virtual entities
